[feat] postService 작성자 확인 기능 추가

+++ src/main/java/com/sparta/dockingfinalproject/post/PostService.java
@@ -11,6 +11,7 @@ import com.sparta.dockingfinalproject.pet.Pet;
 import com.sparta.dockingfinalproject.pet.PetRepository;
 import com.sparta.dockingfinalproject.pet.dto.PetRequestDto;
 import com.sparta.dockingfinalproject.post.dto.PostDetailResponseDto;
+import com.sparta.dockingfinalproject.post.dto.StatusDto;
 import com.sparta.dockingfinalproject.security.UserDetailsImpl;
 import com.sparta.dockingfinalproject.wish.Wish;
 import com.sparta.dockingfinalproject.wish.WishRepository;
@@ -37,7 +38,8 @@ public class PostService {
   private final AlarmRepositoroy alarmRepositoroy;
 
   public PostService(PostRepository postRepository, WishRepository wishRepository,
-      PetRepository petRepository, CommentRepository commentRepository, AlarmRepositoroy alarmRepositoroy) {
+      PetRepository petRepository, CommentRepository commentRepository,
+      AlarmRepositoroy alarmRepositoroy) {
     this.postRepository = postRepository;
     this.wishRepository = wishRepository;
     this.petRepository = petRepository;
@@ -52,36 +54,39 @@ public class PostService {
 
     List<PostDetailResponseDto> postList = new ArrayList<>();
     for (Post post : posts) {
-      PostDetailResponseDto postDetailResponseDto = PostDetailResponseDto.getPostDetailResponseDto(post);
+      PostDetailResponseDto postDetailResponseDto = PostDetailResponseDto.getPostDetailResponseDto(
+          post);
       postList.add(postDetailResponseDto);
     }
 
     Map<String, Object> data = new HashMap<>();
     data.put("postList", postList);
-    data.put("alarmCount", alarmRepositoroy.findAllByUserAndStatusTrueOrderByCreatedAtDesc(userDetails.getUser()));
+    data.put("alarmCount",
+        alarmRepositoroy.findAllByUserAndStatusTrueOrderByCreatedAtDesc(userDetails.getUser()));
 
     return SuccessResult.success(data);
   }
 
   @Transactional
-  public Map<String, Object> getPosts(Long postId, UserDetailsImpl userDetails) {
-    Post findPost = postRepository.findById(postId).orElseThrow(
-        () -> new DockingException(ErrorCode.POST_NOT_FOUND)
-    );
-
-    Optional<Wish> findWish = wishRepository.findAllByUserAndPost(userDetails.getUser(), findPost);
+  public Map<String, Object> getPost(Long postId, UserDetailsImpl userDetails) {
+    Post findPost = bringPost(postId);
+    Long userId = userDetails.getUser().getUserId();
+    Optional<Wish> findWish = null;
 
     boolean heart = false;
-    if (findWish.isPresent()) {
-      heart = true;
+    if (userId != null) {
+      findWish = wishRepository.findAllByUserAndPost(userDetails.getUser(), findPost);
+      if (findWish.isPresent()) {
+        heart = true;
+      }
+      findPost.addViewCount();
     }
-    findPost.addViewCount();
+
     PostDetailResponseDto postResponseDto = PostDetailResponseDto
         .getPostDetailResponseDto(findPost, heart);
 
     //Comment return data 가공하기
     ArrayList<CommentResultDto> commentDtoList = new ArrayList<>();
-    CommentResultDto commentResultDto = new CommentResultDto();
 
     List<CommentResponseDto> commentResponseDto = commentRepository.findAllByPost(findPost);
     for (CommentResponseDto crd : commentResponseDto) {
@@ -91,7 +96,7 @@ public class PostService {
       LocalDateTime modifiedAt = crd.getModifiedAt();
       String nickname = crd.getUser().getNickname();
 
-      commentResultDto = new CommentResultDto(commentId, comment, nickname, createdAt, modifiedAt);
+      CommentResultDto commentResultDto = new CommentResultDto(commentId, comment, nickname, createdAt, modifiedAt);
       commentDtoList.add(commentResultDto);
     }
 
@@ -106,41 +111,97 @@ public class PostService {
   public Map<String, Object> addPost(PetRequestDto petRequestDto, UserDetailsImpl userDetails) {
     Pet pet = new Pet(petRequestDto);
     Pet savePet = petRepository.save(pet);
+    Long userId = userDetails.getUser().getUserId();
 
+    if (userId != null) {
     Post post = new Post(savePet, userDetails.getUser());
     postRepository.save(post);
+    } else {
+      throw new DockingException(ErrorCode.NO_AUTHORIZATION);
+    }
 
     return SuccessResult.success(new ArrayList<>());
   }
 
   @Transactional
-  public Map<String, Object> updatePost(Long postId, PetRequestDto petRequestDto) {
-    Post findPost = postRepository.findById(postId).orElseThrow(
-        () -> new DockingException(ErrorCode.POST_NOT_FOUND)
-    );
+  public Map<String, Object> updatePost(Long postId, PetRequestDto petRequestDto, UserDetailsImpl userDetails) {
+    Post findPost = bringPost(postId);
+    Map<String, String> data = new HashMap<>();
+    Long userId = userDetails.getUser().getUserId();
+    Long writerId = findPost.getUser().getUserId();
 
+    if (userId.equals(writerId)) {
     Pet pet = findPost.getPet();
     pet.update(petRequestDto);
     findPost.addPet(pet);
 
-    Map<String, String> data = new HashMap<>();
     data.put("msg", "수정 완료");
+    } else {
+      throw new DockingException(ErrorCode.NO_AUTHORIZATION);
+    }
     return SuccessResult.success(data);
   }
 
   @Transactional
-  public Map<String, Object> deletePost(Long postId) {
-    Post findPost = postRepository.findById(postId).orElseThrow(
-        () -> new DockingException(ErrorCode.POST_NOT_FOUND)
-    );
-
+  public Map<String, Object> deletePost(Long postId, UserDetailsImpl userDetails) {
+    Post findPost = bringPost(postId);
     Pet findPet = findPost.getPet();
+    Long userId = userDetails.getUser().getUserId();
+    Long writerId = findPost.getUser().getUserId();
 
+    Map<String, Object> data = new HashMap<>();
+
+    if (userId.equals(writerId)) {
     postRepository.deleteById(findPost.getPostId());
     petRepository.deleteById(findPet.getPetId());
 
-    Map<String, Object> data = new HashMap<>();
     data.put("msg", "삭제 완료");
+    } else {
+      throw new DockingException(ErrorCode.NO_AUTHORIZATION);
+    }
     return SuccessResult.success(data);
+  }
+
+  // 보호 상태 변경하기
+  @Transactional
+  public Map<String, Object> updateStatus(Long postId, StatusDto statusDto,
+      UserDetailsImpl userDetails) {
+    String newStatus = statusDto.getNewStatus();
+
+    Long userId = userDetails.getUser().getUserId();
+
+    Post findPost = bringPost(postId);
+    Long writerId = findPost.getUser().getUserId();
+
+    Map<String, String> data = new HashMap<>();
+
+    // 보호 상태 업데이트하기
+    if (userId.equals(writerId)) {
+      Pet findPet = findPost.getPet();
+      String status = findPet.getIsAdopted();
+      String isAdopted = "";
+
+      if (newStatus.equals(status)) {
+        throw new DockingException(ErrorCode.NO_DIFFERENCE);
+      } else if (newStatus.equals("보호종료")) {
+        isAdopted = "보호종료";
+      } else if (newStatus.equals("입양진행중")) {
+        isAdopted = "입양진행중";
+      }
+
+      findPet.updateStatus(isAdopted);
+
+      data.put("msg", "보호상태가 <" + isAdopted + ">(으)로 변경되었습니다.");
+    } else {
+      throw new DockingException(ErrorCode.NO_AUTHORIZATION);
+    }
+    return SuccessResult.success(data);
+  }
+
+  // postId로 해당 Post 가져오기
+  private Post bringPost(Long postId) {
+    return postRepository.findById(postId).orElseThrow(
+        () -> new DockingException(ErrorCode.POST_NOT_FOUND)
+    );
   }
 }
